+++ packages/hoprd/tsconfig.json
@@ -4,7 +4,7 @@
     "outDir": "./lib",
     "rootDir": "./src"
   },
-  "include": ["src/**/*.ts"],
+  "include": ["src/**/*.ts", "src/main.cts"],
   "typedocOptions": {
     "entryPoints": ["src/index.ts"],
     "out": "docs",
