+++ scripts/run-integration-tests-source.sh
@@ -167,7 +167,7 @@ function setup_node() {
   # Set NODE_ENV=development to rebuild hopr-admin next files
   # at runtime. Necessary to start multiple instances of hoprd
   # in parallel
-  DEBUG="hopr*" NODE_ENV=development node packages/hoprd/lib/main.js \
+  DEBUG="hopr*" NODE_ENV=development node packages/hoprd/lib/main.cjs \
     --admin \
     --adminHost "127.0.0.1" \
     --adminPort ${admin_port} \
