+++ scripts/vm-nat-testing/docker-compose-nat.yaml
@@ -26,7 +26,7 @@ services:
       - -c
       - |
         ./prepare-node.sh
-        node /opt/hopr/packages/hoprd/lib/main.js --admin --adminHost 0.0.0.0 \
+        node /opt/hopr/packages/hoprd/lib/main.cjs --admin --adminHost 0.0.0.0 \
           --healthCheck --healthCheckHost 0.0.0.0 --init --rest --restHost 0.0.0.0 --environment hardhat-localhost \
           --apiToken MyT0ken123^ --password switzerland --testPreferLocalAddresses \
           --identity /var/hopr/identities/local-nat-node1.id --testUseWeakCrypto
@@ -47,7 +47,7 @@ services:
       - -c
       - |
         ./prepare-node.sh
-        node /opt/hopr/packages/hoprd/lib/main.js --admin --adminHost 0.0.0.0 \
+        node /opt/hopr/packages/hoprd/lib/main.cjs --admin --adminHost 0.0.0.0 \
           --healthCheck --healthCheckHost 0.0.0.0 --init --rest --restHost 0.0.0.0 --environment hardhat-localhost \
           --apiToken MyT0ken123^ --password switzerland --testPreferLocalAddresses \
           --identity /var/hopr/identities/local-nat-node2.id --testUseWeakCrypto
