+++ scripts/vm-nat-testing/docker-compose-pub.yaml
@@ -13,7 +13,7 @@ services:
       - -c
       - |
         ./prepare-node.sh
-        node /opt/hopr/packages/hoprd/lib/main.js --admin --adminHost 0.0.0.0 \
+        node /opt/hopr/packages/hoprd/lib/main.cjs --admin --adminHost 0.0.0.0 \
           --healthCheck --healthCheckHost 0.0.0.0 --init --rest --restHost 0.0.0.0 --environment hardhat-localhost \
           --apiToken MyT0ken123^ --password switzerland --testPreferLocalAddresses \
           --identity /var/hopr/identities/local-public.id --testUseWeakCrypto --announce
